backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy>=2.0
alembic
psycopg[binary]
passlib[bcrypt]
pyjwt
python-multipart
pydantic>=2
email-validator
pytesseract
opencv-python
pillow
pdfplumber
python-magic
camelot-py[cv]
pandas
PyMuPDF
google-generativeai
